Breakout PointPathType
