+++ src/Svg.Skia/Svg/PathPointType.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Svg.Skia
+{
+    [Flags]
+    internal enum PathPointType : byte
+    {
+        Start = 0,
+        Line = 1,
+        Bezier = 3,
+        Bezier3 = 3,
+        PathTypeMask = 0x7,
+        DashMode = 0x10,
+        PathMarker = 0x20,
+        CloseSubpath = 0x80
+    }
+}
